Langchain/requirement.txt
langchain
python-dotenv
ipykernel
pydantic
langchain_community
pypdf
beautifulsoup4
arxiv
PyMuPDF 
langchain-text-splitters
langchain-deepseek
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain_core
langchain_groq
fastapi
uvicorn
langserve
sse_starlette